Make the ConnectClient compliant with http.Handler
Making the `ConnectClient` structure compliant with the `http.Handler`
allows it to handle directly if wanted the response from an http.Server
instance.
A dedicated interface has been created to hold the answer to RecastAI
server and has be named `MessageWriter`.

Also fixes a potential panic when a request with a nil body is passed to
the `ParseConnectorMessage` function.